Bill screen logic finished, all working

+++ app/src/main/java/com/example/splitbill/data/local/dao/BillDao.kt
@@ -37,6 +37,6 @@ interface BillDao {
     fun getAllBills(): Flow<List<Bill>>
 
     @Query("SELECT * FROM Bill WHERE id = :id LIMIT 1")
-    suspend fun getBillById(id: Int): Bill?
+    suspend fun getBillById(id: Long): Bill?
 
 }
